% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_array.R
\name{conv_array}
\alias{conv_array}
\title{Convert dataset to array}
\usage{
conv_array(data)
}
\arguments{
\item{data}{Dataset to convert in array}
}
\value{
Returns the input dataset in a array
}
\description{
Convert dataset to array
}
\examples{
# Example 1
# Generating some array values with dataframe

data <- data.frame(
  'month' = 1:4,
  'cost' = c(23, 10, 15, 25),
  'number' = c(1.2, 3.4, 4.5, 5.6))
conv_array(data)
}
\author{
Catherine Rincon, \email{catherine.rincon@udea.edu.co}
}
