% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_min_max.R
\name{norm_min_max}
\alias{norm_min_max}
\title{Min-max normalization}
\usage{
norm_min_max(x, inf, scaler = "transform")
}
\arguments{
\item{x}{Dataset to normalize}

\item{inf}{Dataset with range, minimum, and maximum as
min_max_range function output array}

\item{scaler}{The argument to transform with min-max normalization
or inverse of normalization in real values}
}
\value{
Returns two options:
with scaler parameter in transform, returns input data with
min-max normalization
with scaler parameter in inverse, returns data with real values
}
\description{
Min-max normalization
}
\examples{
# Example 1
# Generating dataset, use the min_max_range function to obtain range and
# use the norm_min_max function with scaler argument equal to transform

data <- matrix(c(25, 30, 40, 70, 20, 10), nrow=2, ncol=3)
inf <- min_max_range(data)
norm_min_max(data, inf, scaler='transform')
}
\author{
Catherine Rincon, \email{catherine.rincon@udea.edu.co}
}
