% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_CNN.R
\name{param_CNN}
\alias{param_CNN}
\title{Define the hyperparameters of the convolutional neural network CNN}
\usage{
param_CNN(n_steps, n_filter, size_filter, pool_size, n_neuron)
}
\arguments{
\item{n_steps}{Number of steps to split into samples CNN}

\item{n_filter}{Number of filters in convolutional layer}

\item{size_filter}{Size of filter in convolutional layer}

\item{pool_size}{Size of pool in max_pooling layer}

\item{n_neuron}{Number of neurons in dense layer}
}
\value{
Return CNN model architecture and model summary
}
\description{
Define the hyperparameters of the convolutional neural network CNN
}
\examples{
# Example 1
# Generating a new CNN arquitecture

\dontrun{
library(keras)

model_CNN <- param_CNN(n_steps=6, n_filter=4,size_filter=2, pool_size=2, n_neuron=5)

summary(model_CNN)

}
}
\author{
Catherine Rincon, \email{catherine.rincon@udea.edu.co}
}
