% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_LSTM.R
\name{param_LSTM}
\alias{param_LSTM}
\title{Define the hyperparameters of the long-short term memory neural network LSTM}
\usage{
param_LSTM(n_steps_LSTM, n_features, f_n_neuron, f_drop, s_n_neuron, s_drop)
}
\arguments{
\item{n_steps_LSTM}{Number of steps to split into samples LSTM}

\item{n_features}{LSTM training features dataset number}

\item{f_n_neuron}{Number of neurons of first LSTM layer}

\item{f_drop}{First percent dropout rate}

\item{s_n_neuron}{Number of neurons of second LSTM layer}

\item{s_drop}{Second percent dropout rate}
}
\value{
Return LSTM model architecture and model summary
}
\description{
Define the hyperparameters of the long-short term memory neural network LSTM
}
