% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_unv.R
\name{split_unv}
\alias{split_unv}
\title{Split a univariate dataset into samples}
\usage{
split_unv(dataset, n_steps)
}
\arguments{
\item{dataset}{Dataset to split into samples as univariate dataset}

\item{n_steps}{Number (integer) of steps to split into samples}
}
\value{
Returns two data sets: one with the information of the x (features)
and another with the information of the y (target)
}
\description{
Split a univariate dataset into samples
}
\examples{
# Example 1
# Generating dataset to split into samples as univariate dataset

data <- matrix(c(25, 30, 40, 70, 20, 10), nrow=6, ncol=1)
split_unv(data, n_steps = 3)

# Example 2
# Generating dataset with data NASA, split the temperature with split_unv
# function to obtain samples as univariate dataset

data(data_NASA)
temperature <- data_NASA$temperature_measured
split_unv(temperature, n_steps = 3)
}
\author{
Catherine Rincon, \email{catherine.rincon@udea.edu.co}
}
