% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_tuning_CNN.R
\name{hp_tuning_CNN}
\alias{hp_tuning_CNN}
\title{Hyperparameter tuning CNN}
\usage{
hp_tuning_CNN(data, n_steps, n_filter, neuron, learn_rate, epoc)
}
\arguments{
\item{data}{Dataset to split and tuning}

\item{n_steps}{Number of steps to split into samples}

\item{n_filter}{Number of filters in the convolutional layer}

\item{neuron}{Number of neuron in the dense layer}

\item{learn_rate}{Learning rate in the compilation with Adam optimizer}

\item{epoc}{Number of epochs in the training model}
}
\value{
Returns:
Dataframe with the combinations of the parameters and the calculated error
Better combination of hyperparameters that minimizes the train and test error
}
\description{
Hyperparameter tuning CNN
}
\author{
Catherine Rincon, \email{catherine.rincon@udea.edu.co}
}
