% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_range.R
\name{min_max_range}
\alias{min_max_range}
\title{Calculate the range of a dataset}
\usage{
min_max_range(x)
}
\arguments{
\item{x}{Dataset to calculate the minimum and maximum}
}
\value{
Returns a matrix with 2 rows and p columns (variables) with
the range (minimum and maximum) of dataset
}
\description{
Calculate the range of a dataset
}
\examples{
# Example 1
# Generating dataset to obtain range: minimum and maximum

data <- matrix(c(25, 30, 40, 70, 20, 10), nrow=2, ncol=3)
min_max_range(data)
}
\author{
Catherine Rincon, \email{catherine.rincon@udea.edu.co}
}
