% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_seq.R
\name{split_seq}
\alias{split_seq}
\title{Split a multivariate dataset sequence into samples}
\usage{
split_seq(dataset, n_steps)
}
\arguments{
\item{dataset}{Dataset to split into samples as multivariate sequences dataset}

\item{n_steps}{Number of steps to split into samples}
}
\value{
Returns two data sets: one with the information of the x
and another with the information of the y
}
\description{
Split a multivariate dataset sequence into samples
}
\author{
Catherine Rincon, \email{catherine.rincon@udea.edu.co}
}
